CHANGED: Handling of ERefFrame enumeration
Up till now there was some code duplication with regard to ERefFrame enumeration.
One version was present in the TRefSystemFactory, while other was being used
in the TDataParameters.

After refactoring, the only ERefFrame enumeration remaining is in the
TRefSystemFactory. The order of the enum values has been changed to clearly represent
whether it is a global reference frame, a projection or an MLA.
TDataParameters was changed to use this enumeration.

New class TRefFrameInfo that provides information about the given ERefFrame was added.
For now, it provides the functionality to:
   - check whether a given number can be converted to ERefFrame enum value
   - safely (exception thrown on error) convert a number to ERefFrame enum value
   - enumerate over all the possible ERefFrame enum values
   - retrieve user friendly name (string) for the ERefFrame
   - retrieve the default coordinate system for the ERefFrame
   - check whether the ERefFrame is local (true for MLA)
   - provide a single method for creating the TAReferenceFrame (by using
     the TRefSystemFactory methods) from a given ERefFrame (and origin
     if needed)