[SurveyLib_LGC] ADDED: Renamed the typedef for the 'real' datatype - to TReal. real was conflicting with standard library function named real.
In order to simplify and to allow for a simple library replacement a new class TMatrixImpl was created. It will provide the Eigen matrix operations to the TMatrix class. Such approach allows to limit the Eigen related code to a single class/file (other classes will be using TMatrix as if nothing had changed).
Simplification of the code TMatrix.h and TMatrix.cpp.